+++ apps/fluent-tester/src/FluentTester/TestComponents/Menu/MenuTest.tsx
@@ -14,6 +14,7 @@ import { Stack } from '@fluentui-react-native/stack';
 import { stackStyle } from '../Common/styles';
 import { MENU_TESTPAGE } from './consts';
 import { Test, TestSection, PlatformStatus } from '../Test';
+import { Text } from '@fluentui-react-native/experimental-text';
 import { E2EMenuTest } from './E2EMenuTest';
 
 const MenuDefault: React.FunctionComponent = () => {
@@ -38,7 +39,7 @@ const MenuDefault: React.FunctionComponent = () => {
 const MenuCheckmarks: React.FunctionComponent = () => {
   return (
     <Stack style={stackStyle}>
-      <Menu defaultChecked={{ itemOne: true }}>
+      <Menu defaultChecked={['itemOne']}>
         <MenuTrigger>
           <Button>All checkmark items</Button>
         </MenuTrigger>
@@ -50,7 +51,7 @@ const MenuCheckmarks: React.FunctionComponent = () => {
           </MenuList>
         </MenuPopover>
       </Menu>
-      <Menu hasCheckmarks checked={{ itemTwo: true }}>
+      <Menu hasCheckmarks checked={['itemTwo']}>
         <MenuTrigger>
           <Button>Some controlled checkmark items with alignment</Button>
         </MenuTrigger>
@@ -80,17 +81,27 @@ const MenuCheckmarks: React.FunctionComponent = () => {
 };
 
 const MenuRadioItem: React.FunctionComponent = () => {
+  const [checked, setChecked] = React.useState(['itemOne']);
+  const onCheckedChange = React.useCallback(
+    (_e, checkedChange) => {
+      setChecked(checkedChange);
+    },
+    [setChecked],
+  );
+
   return (
     <Stack style={stackStyle}>
-      <Menu defaultChecked={{ itemOne: true }}>
+      <Text>Current checked: {checked.join(' ')}</Text>
+      <Menu defaultChecked={['itemOne']} onCheckedChange={onCheckedChange}>
         <MenuTrigger>
           <Button>Items with radio selection</Button>
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItemRadio name="itemOne" content="A MenuItem with checkmark" />
-            <MenuItemRadio name="itemTwo" content="Another MenuItem with checkmark" />
-            <MenuItemRadio name="itemThree" content="A third MenuItem with checkmark" />
+            <MenuItemRadio name="itemOne" content="A MenuItem with checkmark and radio selection" />
+            <MenuItemRadio name="itemTwo" content="Another MenuItem with checkmark and radio selection" />
+            <MenuItemRadio name="itemThree" content="A third MenuItem with checkmark and radio selection" />
+            <MenuItemCheckbox name="itemFour" content="A MenuItem with checkmark and toggle selection" />
           </MenuList>
         </MenuPopover>
       </Menu>