+++ packages/experimental/Menu/src/Menu/Menu.types.ts
@@ -16,5 +16,6 @@ export interface MenuState extends MenuProps {
   isSubmenu: boolean;
   parentPopoverHoverOutTimer?: NodeJS.Timeout;
   setOpen: (e: InteractionEvent, isOpen: boolean, bubble?: boolean) => void;
+  shouldFocusOnContainer: boolean;
   triggerRef: React.RefObject<React.Component>;
 }