+++ packages/experimental/Menu/src/Menu/useMenu.ts
@@ -1,5 +1,6 @@
-import { InteractionEvent } from '@fluentui-react-native/interactive-hooks';
+import { InteractionEvent, isMouseEvent } from '@fluentui-react-native/interactive-hooks';
 import React from 'react';
+import { Platform } from 'react-native';
 import { useMenuContext } from '../context/menuContext';
 import { MenuProps, MenuState } from './Menu.types';
 
@@ -14,8 +15,10 @@ export const useMenu = (props: MenuProps): MenuState => {
   const triggerRef = React.useRef();
   const context = useMenuContext();
   const isSubmenu = context.triggerRef !== null;
-  const isControlled = typeof props.open !== 'undefined';
-  const [open, setOpen] = useMenuOpenState(isControlled, props, context.setOpen);
+  const isOpenControlled = typeof props.open !== 'undefined';
+  const [open, shouldFocusOnContainer, setOpen] = useMenuOpenState(isOpenControlled, props, context.setOpen);
+
+  const [checked, onCheckedChange] = useMenuCheckedState(props);
 
   // Default behavior for submenu is to open on hover
   // the ...props line below will override this behavior for a submenu
@@ -33,9 +36,12 @@ export const useMenu = (props: MenuProps): MenuState => {
     ...props,
     open,
     setOpen,
+    shouldFocusOnContainer,
+    checked,
+    onCheckedChange,
     triggerRef,
     isSubmenu,
-    isControlled,
+    isControlled: isOpenControlled,
     parentPopoverHoverOutTimer,
   };
 };
@@ -44,25 +50,30 @@ const useMenuOpenState = (
   isControlled: boolean,
   props: MenuProps,
   parentSetOpen: (e: InteractionEvent, isOpen: boolean, bubble?: boolean) => void,
-): [boolean, (e: InteractionEvent, isOpen: boolean, bubble?: boolean) => void] => {
+): [boolean, boolean, (e: InteractionEvent, isOpen: boolean, bubble?: boolean) => void] => {
   const { defaultOpen, onOpenChange, open } = props;
   const initialState = typeof defaultOpen !== 'undefined' ? defaultOpen : !!open;
   const [openInternal, setOpenInternal] = React.useState<boolean>(initialState);
+  const [shouldFocusOnContainer, setShouldFocusOnContainer] = React.useState<boolean>(false);
 
   const state = isControlled ? open : openInternal;
 
   const setOpen = React.useCallback(
     (e: InteractionEvent, isOpen: boolean, bubble?: boolean) => {
       const openPrev = state;
-      if (!isControlled) {
-        if (!isOpen || lastCloseTimestamp + delayOpen <= Date.now()) {
-          setOpenInternal(isOpen);
-        }
-
-        if (!isOpen) {
-          lastCloseTimestamp = Date.now();
-        }
+      if (!isControlled && (!isOpen || lastCloseTimestamp + delayOpen <= Date.now())) {
+        setOpenInternal(isOpen);
       }
+
+      if (isOpen && Platform.OS === ('win32' as any) && isMouseEvent(e)) {
+        setShouldFocusOnContainer(true);
+      }
+
+      if (!isOpen) {
+        setShouldFocusOnContainer(false);
+        lastCloseTimestamp = Date.now();
+      }
+
       if (onOpenChange && openPrev !== isOpen) {
         onOpenChange(e, isOpen);
       }
@@ -74,5 +85,26 @@ const useMenuOpenState = (
     [isControlled, state, onOpenChange, setOpenInternal, parentSetOpen],
   );
 
-  return [state, setOpen];
+  return [state, shouldFocusOnContainer, setOpen];
+};
+
+const useMenuCheckedState = (props: MenuProps): [string[], (e: InteractionEvent, checked: string[]) => void] => {
+  const { checked, defaultChecked, onCheckedChange: onCheckedChangeOriginal } = props;
+  const [checkedInternal, setCheckedInternal] = React.useState(defaultChecked ?? checked ?? []);
+
+  const isControlled = typeof checked !== 'undefined';
+  const state = isControlled ? checked : checkedInternal;
+
+  const onCheckedChange = React.useCallback(
+    (e: InteractionEvent, checked: string[]) => {
+      if (!isControlled) {
+        setCheckedInternal(checked);
+      }
+
+      onCheckedChangeOriginal?.(e, checked);
+    },
+    [isControlled, setCheckedInternal, onCheckedChangeOriginal],
+  );
+
+  return [state, onCheckedChange];
 };