+++ packages/experimental/Menu/src/MenuList/MenuList.tsx
@@ -39,7 +39,7 @@ export const MenuList = compose<MenuListType>({
   useRender: (userProps: MenuListProps, useSlots: UseSlots<MenuListType>) => {
     const menuList = useMenuList(userProps);
     const contextValue = useMenuListContextValue(menuList);
-    const Slots = useSlots(menuList);
+    const Slots = useSlots(menuList.props);
 
     return (_final: MenuListProps, children: React.ReactNode) => {
       // macOS needs an extra FocusZone slot to get proper keyboarding behavior with menus.