+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Tue, 07 Jun 2022 22:59:24 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Jun 2022 23:58:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.29.15
+
+Mon, 13 Jun 2022 23:58:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.22.28
+- Bump @fluentui-react-native/callout to v0.20.7
+- Bump @fluentui-react-native/checkbox to v0.15.12
+- Bump @fluentui-react-native/contextual-menu to v0.16.36
+- Bump @fluentui-react-native/focus-trap-zone to v0.8.45
+- Bump @fluentui-react-native/focus-zone to v0.9.32
+- Bump @fluentui-react-native/interactive-hooks to v0.16.3
+- Bump @fluentui-react-native/link to v0.15.46
+- Bump @fluentui-react-native/menu-button to v0.7.68
+- Bump @fluentui-react-native/pressable to v0.8.45
+- Bump @fluentui-react-native/radio-group to v0.14.52
+- Bump @fluentui-react-native/tabs to v0.6.51
 
 ## 0.29.14